feat: Add Author Item to Table to View Author
Added author th tag to question_list.html.
We then added a text-center class to the tr element so that th elements are aligned in the middle, and we also specified style="width:50%" so that title width accounts for 50% of the total.

Displayed the question writer by inserting <td>{{question.user.username}}</td> in the for statement.
Then, I added a text-center class to the tr element so that the contents of the table are aligned in the middle, and added a text-start class so that the title is aligned to the left.