feat: Add Author View Area to Question Details
The design such as margins and alignment has also been slightly changed using bootstrap to show the author and the date of creation in question_detail.html.

The author was added to the answer.
Similar to the question, the author is displayed immediately above the date and time of creation.